still a ton to do
Basically want to store each and every object we get from Discord with a `_shardId` attached to it so we can clear the cache more easily when that one shard goes down.

The mock gateway will just echo every payload we receive from discord, and will also each all our sent payloads to discord. You just have to replace the gateway url with mock's websocket url. This allows for large bots to not have to re-identify since all the data is in MongoDB. Should allow for easy integration with detritus-client.